+++ Changes.md
@@ -1,7 +1,10 @@
 Version Information
 ===================
+Version 3.3.0.1.12
+  1. Added 'Customise individual blocks' to the 'Colours' settings tab - not complete, work in progress.
+
 Version 3.3.0.1.11
-  1 Added main background, nav-drawer background, block header background and text colour settings.
+  1. Added main background, nav-drawer background, block header background and text colour settings.
 
 Version 3.3.0.1.10
   1. Continuation of 'User institution specific header logo' development.